Moving ExceptionController to controller package
