+++ grails-app/controllers/com/miniasaaslw/controller/BaseController.groovy
@@ -1,6 +1,6 @@
 package com.miniasaaslw.controller
 
-import com.miniasaaslw.ExceptionController
+
 import com.miniasaaslw.domain.customer.Customer
 import com.miniasaaslw.domain.security.User
